+++ environment-config.js
@@ -1,7 +1,9 @@
 (() => {
   const config = Object.freeze({
     name: 'production', label: '',
+    dataBackend: 'google_sheets',
     backendUrl: 'https://script.google.com/macros/s/AKfycbw_c_AGjrq91Gjl_UrYMDxIzQtSUsW0PPm6H3WmJUEZZN2uXmvrXbKTlVxahLZ6OOzq/exec',
+    postgresApiUrl: '',
     storagePrefix: '', serviceWorkerUrl: './service-worker.js'
   });
   const storageKey = key => `${config.storagePrefix}${key}`;